Refactor network monitoring and memory types
This commit includes significant changes to improve network monitoring and adjust
memory types:

- Convert memory values from u64 to f32 for better precision
- Restructure network monitoring with separate RX/TX stats
- Add IP address tracking for network interfaces
- Improve memory size formatting and conversion
- Update UI to display detailed network statistics

+++ system_monitor/src/models/memory.rs
@@ -1,23 +1,23 @@
 
 #[derive(Debug, Clone)]
 pub struct MemoryInfo {
-    pub ram_used: u64,
-    pub ram_total: u64,
-    pub swap_used: u64,
-    pub swap_total: u64,
-    pub disk_used: u64,
-    pub disk_total: u64,
+    pub ram_used: f32,
+    pub ram_total: f32,
+    pub swap_used: f32,
+    pub swap_total: f32,
+    pub disk_used: f32,
+    pub disk_total: f32,
 }
 
 impl Default for MemoryInfo {
     fn default() -> Self {
         MemoryInfo {
-            ram_used: 0,
-            ram_total: 0,
-            swap_used: 0,
-            swap_total: 0,
-            disk_used: 0,
-            disk_total: 0,
+            ram_used: 0.0,
+            ram_total: 0.0,
+            swap_used: 0.0,
+            swap_total: 0.0,
+            disk_used: 0.0,
+            disk_total: 0.0,
         }
     }
 }
